// src/app/models/mant.interface.ts
export interface ResponseApi {
  CAMPANA:    string;
  RESULTADO:  string;
  TELEFONO:   string;
}


export interface LeadDto {
  id:               string;
  campana:          string;
  cod_proveedor:     string;
  fecha_captacion:   string;
  nombre:           string;
  ape1:             string;
  ape2:             string;
  nif:              string;
  telefono:         string;
  email:            string;
  direccion:        string;
  codigo_postal:     string;
  poblacion:        string;
  provincia:        string;
  acepta1:          string;
  acepta2:          string;
  acepta3:          string;
  num1:             string;
  num2:             string;
  num3:             string;
  dual1:            string;
  dual2:            string;
  dual3:            string;
  variable1:        string;
  variable2:        string;
  variable3:        string;
  memo:             string;
  fecha:            string;
  hora:             string;
  foto1:            string;
  foto2:            string;
  comercial:        string;
  centro:           string;
}